Refactor utilities, add a database utility module.
Add a new utility module for DB related
functions. This is to be used by test and data analysis code.

Move domain_utils.py to automation/utilities.
Remove repeating PSL related code.
