+++ test/utilities.py
@@ -1,15 +1,10 @@
-from publicsuffix import PublicSuffixList, fetch
 import SimpleHTTPServer
 import SocketServer
 import threading
-import codecs
-import plyvel
 import os
-import sqlite3
 from random import choice
 from os.path import realpath, dirname
 LOCAL_WEBSERVER_PORT = 8000
-PSL_CACHE_LOC = '/tmp/public_suffix_list.dat'
 BASE_TEST_URL_DOMAIN = "localtest.me"
 BASE_TEST_URL_NOPATH = "http://%s:%s" % (BASE_TEST_URL_DOMAIN,
                                          LOCAL_WEBSERVER_PORT)
@@ -43,65 +38,3 @@ def rand_str(size=8):
     """Return random string with the given size."""
     RAND_CHARS = "abcdefghijklmnopqrstuvwxyz0123456789"
     return ''.join(choice(RAND_CHARS) for _ in range(size))
-
-
-def get_psl():
-    """
-    Grabs an updated public suffix list.
-    """
-    if not os.path.isfile(PSL_CACHE_LOC):
-        print "%s does not exist, downloading a copy." % PSL_CACHE_LOC
-        psl_file = fetch()
-        with codecs.open(PSL_CACHE_LOC, 'w', encoding='utf8') as f:
-            f.write(psl_file.read())
-    psl_cache = codecs.open(PSL_CACHE_LOC, encoding='utf8')
-    return PublicSuffixList(psl_cache)
-
-
-def query_db(db, query, params=None):
-    """Run a query against the given db.
-
-    If params is not None, securely construct a query from the given
-    query string and params.
-    """
-    with sqlite3.connect(db) as con:
-        if params is None:
-            rows = con.execute(query).fetchall()
-        else:
-            rows = con.execute(query, params).fetchall()
-    return rows
-
-
-def get_javascript_content(data_directory):
-    """Yield key, value pairs from the deduplicated leveldb content database
-
-    Parameters
-    ----------
-    data_directory : str
-        root directory of the crawl files containing `javascript.ldb`
-    """
-    db_path = os.path.join(data_directory, 'javascript.ldb')
-    db = plyvel.DB(db_path,
-            create_if_missing = False,
-            compression = 'snappy')
-    for content_hash, content in db.iterator():
-        yield content_hash, content
-    db.close()
-
-
-def get_javascript_entries(db, all_columns=False):
-    if all_columns:
-        select_columns = "*"
-    else:
-        select_columns = "script_url, symbol, operation, value, parameter_index,\
-             parameter_value"
-
-    return query_db(db, "SELECT %s FROM javascript" % select_columns)
-
-def any_command_failed(db):
-    """Returns True if any command in a given database failed"""
-    rows = query_db(db, "SELECT * FROM CrawlHistory;")
-    for row in rows:
-        if row[3] != 1:
-            return True
-    return False
